feat: 1단계 - @MVC 구현하기 기능 요구사항 1 구현
